+++ src/main/java/cn/zjiali/robot/model/response/CsylqResponse.java
@@ -1,4 +1,4 @@
-package cn.zjiali.robot.entity.response;
+package cn.zjiali.robot.model.response;
 
 import com.google.gson.annotations.SerializedName;
