+++ src/main/java/cn/zjiali/robot/model/response/FortuneResponse.java
@@ -1,4 +1,4 @@
-package cn.zjiali.robot.entity.response;
+package cn.zjiali.robot.model.response;
 
 import lombok.Data;
