+++ src/main/java/cn/zjiali/robot/model/response/GylqResponse.java
@@ -1,29 +1,29 @@
-package cn.zjiali.robot.entity.response;
+package cn.zjiali.robot.model.response;
+
 
 import com.google.gson.annotations.SerializedName;
 
 /**
  * @author zJiaLi
- * @since 2021-04-08 18:44
+ * @since 2021-04-08 18:38
  */
-public class YllqResponse {
+public class GylqResponse {
 
-    @SerializedName("number1")
+    @SerializedName(value = "number1")
     private String number1;
-    @SerializedName("number2")
+    @SerializedName(value = "number2")
     private String number2;
-    @SerializedName("haohua")
+    @SerializedName(value = "haohua")
     private String haohua;
-    @SerializedName("shiyi")
+    @SerializedName(value = "qianyu")
+    private String qianyu;
+    @SerializedName(value = "shiyi")
     private String shiyi;
-    @SerializedName("jieqian")
+    @SerializedName(value = "jieqian")
     private String jieqian;
-    @SerializedName("zhushi")
-    private String zhushi;
-    @SerializedName("baihua")
-    private String baihua;
-    @SerializedName("type")
+    @SerializedName(value = "type")
     private String type;
+
 
     public String getNumber1() {
         return number1;
@@ -49,6 +49,14 @@ public class YllqResponse {
         this.haohua = haohua;
     }
 
+    public String getQianyu() {
+        return qianyu;
+    }
+
+    public void setQianyu(String qianyu) {
+        this.qianyu = qianyu;
+    }
+
     public String getShiyi() {
         return shiyi;
     }
@@ -63,22 +71,6 @@ public class YllqResponse {
 
     public void setJieqian(String jieqian) {
         this.jieqian = jieqian;
-    }
-
-    public String getZhushi() {
-        return zhushi;
-    }
-
-    public void setZhushi(String zhushi) {
-        this.zhushi = zhushi;
-    }
-
-    public String getBaihua() {
-        return baihua;
-    }
-
-    public void setBaihua(String baihua) {
-        this.baihua = baihua;
     }
 
     public String getType() {
