+++ src/main/java/cn/zjiali/robot/model/response/SignInDataResponse.java
@@ -1,4 +1,4 @@
-package cn.zjiali.robot.entity.response;
+package cn.zjiali.robot.model.response;
 
 
 import lombok.Data;
@@ -50,7 +50,7 @@ public class SignInDataResponse {
         /**
          * 当前等级
          *
-         * @note 2020/10/31 修改
+         *  2020/10/31 修改
          */
         private String currentLevel;
