+++ src/main/java/cn/zjiali/robot/model/response/TodayOnHistoryResponse.java
@@ -1,4 +1,4 @@
-package cn.zjiali.robot.entity.response;
+package cn.zjiali.robot.model.response;
 
 /**
  * 历史上的今天
